NFP-class

An S4 object for storing network fingerprint similarity score information.

Slots

raw_score,

a numeric vector, network fingerprint based on reference networks before standardization.

randomized_score,

a data frame, the permulated similarity score.

standardized_score,

a numeric vector, the final standardized network fingerprint.

cluster,

an APResult list, more details see package **apcluster**, each element provides a cluster information of a biological network based on one reference networks. #'@section method:

perm_score, signature(object = "NFP"): extract the randomized similarity score

cluster_info, signature(object = "NFP"): extract the cluster information

sub_NFP, signature(object = "NFP"): subset of NFP object

plot, signature(object, type = "character", p_size = "numeric", l_size = 'numeric'): plot NFP results

show, signature(object = "NFP"): display methods for S4 classes NFP, see also show

See also

show-methods, plot-methods, perm_score-methods, cluster_info-methods, sub_NFP-methods