NFPRefnet-class

An S4 object for storing NFP reference network information.

Slots

Refnet,

object of graphNEL list represents the basic networks, and each elements contains a group of basic networks.

group,

a character vector whose length is the same with Refnet, the group names of basic networks.

name,

names of the basic networks, with the same data structure with Refnet.

organism,

character, indicating the activation organism of basic networks. #'@section method:

net, signature(object = "NFPRefnet"): extract the basic networks

group, signature(object = "NFPRefnet"): extract group information

subnet, signature(object = "NFPRefnet"): subset basic networks, e.g. a group of a networks or same networks of a given group

refnet_name, signature(object = "NFPRefnet"): the names of basic networks

show, signature(object = "NFPRefnet"): display methods for S4 classes NFPRefnet, see also show

See also

show-methods, net-methods, refnet_name-methods, group-methods, subnet-methods